Fix missing elements, add gui elements and string interpolation
I noticed that when calling for :GetElements(), it returns an empty table, leaving me unable to modify any elements such as rigs to be replaced by a clone of the player's character. I fixed that and modified the path for MoonAnimator's CoreGui, leading to ReplicatedStorage instead.

There, you just have to make a ScreenGui with similarly named elements and parent the ScreenGui to the PlayerGui before playing the cutscene in order to make it work (with the exception of the Letter Box). I also added String interpolation for the Subtitles to work.